+++ src/library/DIPS.Mobile.UI/Resources/Icons/iOS/Icons.cs
@@ -1,12 +1,9 @@
+using DIPS.Mobile.UI.API.Library;
 using UIKit;
 
 namespace DIPS.Mobile.UI.Resources.Icons;
 
 public static partial class Icons
 {
-    public static bool TryGetUIImage(IconName iconName, out UIImage? uiImage)
-    {
-        uiImage = UIImage.FromBundle(GetIconName(iconName));
-        return uiImage != null;
-    }
+    public static bool TryGetUIImage(IconName iconName, out UIImage? uiImage) => DUI.TryGetUIImageFromBundle(GetIconName(iconName), out uiImage);
 }